Avoid ambiguity with "may not" and a "MAY" requirement.
RFC 2119 defines "MAY" as a keyword, which may cause confusion with "may
not", even though the latter is not uppercased.

This hopefully better clarifies that the server in question may or may
not be an origin server, as opposed to implying that origin servers are
prevented from returning a 451 response status.