Rename aesgcm128 to aesgcm to provide an update path
In order to allow existing applications to update from one octet of
padding per record to two (PR #136, Issue #135), this PR renames
"aesgcm128" to "aesgcm" as the value for the Content-Encoding header,
as well as the parameter name in the Crypto-Key header.

This also affects the context used for deriving the nonce and the
content encryption key. Examples have been updated.

What I did not update is the JWE mapping (currently "A128GCM"). Do
we wish to change that to "AGCM", which may be less descriptive?