cookie-alone: add path-match restriction.
This patch adds a path-match constraint to the storage algorithm for
non-secure cookies. This ensures that non-secure cookies cannot overlay
secure cookies for a given path, but allows secure and non-secure
cookies with the same name to exist on distinct paths.

Based on Chrome's implementation experience, the lack of such a
constraint breaks sites in the wild (Blizzard's forums, for example,
as documented in https://crbug.com/580770).

Addresses portions of httpwg/http-extensions#223.