You can't actually receive a 425 if the handshake doesn't complete
@mcmanus was confused by the text here and I confess that I don't remember why
it was added.  There are actually esoteric cases where you get server data that
you can read before the handshake is complete (QUIC tends to enable that sort
of mess), but there is no circumstance where the client can read the 425 but
can't also determine whether that the handshake is not done to its
satisfaction.  The client might decide that it wants to abort and not provide a
certificate or something, but most implementations don't really allow that sort
of messing around.