Manage out of order early data delivery
The best and easiest plan is to discard, but we know that throwing out
perfectly good packets can have bad effects on performance.  It sends a bad
signal to a congestion controller for instance.  So we need to say either
discard, or ensure that the data is properly attributed and consistently
processed.

Closes #391.