Some suggestions for forward compatibility
Some of this is rewording.

There are two new paragraphs talking about extensions: one deals with dictionaries, the other deals
with extensions to a SH definition.

I also removed the "starts with 'h'" example for greasing.  We know that X- lead to all sorts of
mess, so choosing a bad example here could be just as bad.  I'm happy to split that out or be talked
into backing it out, of course.