[RFC2665bis] Fix nameless/valueless serialization.
Nameless cookies serialize as `value`. Valueless cookies serialize as
`name=`. No, this isn't what we would have chosen if we were designing
cookies from scratch, but it's Chromium and Firefox's shipping behavior,
and it's reasonable to just accept.

This specification change is covered by 0021 and 0022 in the WPT
repository (see
https://wpt.fyi/results/cookies/http-state/general-tests.html?label=master&label=experimental&aligned).

Closes https://github.com/httpwg/http-extensions/issues/1081.