Big numbers for H3 frame type
Strawman proposal.

In HTTP/3 we have a huge space of possible frame types, so lets use
them during draft development. The values in this commit were picked
by generating a set of GREASE values, selecting one and adding 1
and 2 to it.

The HTTP/2 frame type remains as 0xF because it is only 8-bits,
meaning there is less room to play and greater chance of polluting a
small range. However, I also expect less churn in H2, partly because
I anticipate that implementers of the draft will focus on HTTP/3.

Since we intend to reserve the values 0xF and 0x10 in the final RFC,
I've left the IANA consideration unchanged. I have reserved the
selected experimental values in the temporary IANA registry at
https://github.com/quicwg/base-drafts/wiki/Temporary-IANA-Registry.