Default urgency is 3
When we discussed removing the semantics of urgency levels, we also
thought it would be a good idea to change the default from 1 to 3.

We agreed to do this and landed a change that removed the semantic
text but unfortunately forgot to also change the default.

This change simple fixes the oversight.