Better explain PRIORITY_UPDATE buffering and application
The old text had a loose definition of when to apply buffered frames.
It also had a fair amount of repetition.

This change tweaks the sections a bit so that we have one that discusses
how a server should handle PRIORITY_UPDATES for either initial signal or
reprioritization. It incorporates the suggestions on issue #1260 mixed
in with some text we already had.