refactor priority parameters and Priority header bits
Fixes #1716 and #1720.

We define priority parameters as a structured fields dictionary. But
don't explicitly come out and say that the Priority header fields should
be treated that way. So this change fixes that.

The text in the Priority header section attempts to give a concise
overview of how it is used but that caused confusion. Since the text
just summarised information presented elsewhere, this change refactors
things to point to those sections.

While we are tweaking the priority header section, we can pull the text
about parsing failure handling up into the priority parameters section
so that there is a single place that gathers parsing considerations.

Finally, I noticed that our header registration was a bit old fashioned
in light of what's happening in HTTP core. So I updated that too.