Standardize maximum Expires/Max-Age
The current spec is ambiguous on (1) what the maximum `Expires`/`Max-Age` attribute values can be and (2) whether the two must be consistent. This resolves both by requiring:

(1) The maximum attribute value to be 400 days in the future or less
(2) The maximum set for Max-Age and Expires be consistent.

Why 400 days? The goal was to get close to 13 months so that functions one might perform annually (e.g., selecting insurance benefits for the next year) would work even as specific dates varied slightly. Since Expires deals in a specific date while Max-Age deals in delta-seconds, it seems ideal to select a value with an unambiguous meaning. If the cap were set at 13 months (instead of 400 days) in the future, the maximum TTL varies by the month it was set in and whether or not it's a leap year.

closes #1600