Keep XML files
Generally, I don't think that it is a good idea to specify the use of `.SECONDARY` in makefiles as
it leaves a bunch of intermediate stuff in the repository.  It also means that you have to update
the .gitignore when you add new non-XML files.

In this case, however, the linting step in this repository ends up being very expensive and noisy as
it uses XML input.  Keeping the XML around manages that a little better.