Why truncation
Basically, it saves the effort of writing stuff that you know to be empty.  It's
not a critical feature, but it turns out to be a simple way to deal with not
sending trailers or a body.  In the extreme, you can also drop headers to
produce the best ever HTTP message: 0x0140c8.

CLoses #1946.