message-signatures: fix Content-Digest sha-512 digest on response
I'm not sure what the current digest is, but it doesn't match the 23-byte HTTP body of that message.
We can verify the new digest with:

$ echo -n '{"message": "good dog"}' | sha512sum | cut -d ' ' -f 1 | xxd -r -p | base64
mEWXIS7MaLRuGgxOBdODa3xqM1XdEvxoYhvlCFJ41QJgJc4GTsPp29l5oGX69wWdXymyU0rjJuahq4l5aGgfLQ==

(also verified in a go implementation of message signatures)

As the digest changed, the Signature for the ECDSA P-256 example also changed, I've included a new
signature over the updated digest that verifies based on the key in the spec.