Message contents has newlines, ensure digest consistency
By virtue of the presentation formats available to RFCs, it was ambigous
to readers whether example's content included a trailing LF or not. This was bad
because the content (or even representation) is required for digest
calculation.

This change makes it clear, where needed, what the object bytes are that
we are sticking into the digest calculation. And since that now includes
LF, we need to recompute all the example digest values.

Fixes #2388