Resumable Upload: Offset must not decrease
As discussed in https://github.com/httpwg/http-extensions/issues/2695, this requires the upload offset to not decrease. The server is not allowed to "forget" data appended to an upload. The benefit is that it allows a client to reduce the amount of data buffered while the upload is ongoing as it receives acknowledgements from the server that data has been stored permanently and will not have to be retransmitted.

Closes https://github.com/httpwg/http-extensions/issues/2695