no-vary-search: Allow more efficient caches by assuming consistency
For most known use cases, a particular resource is consistently
invariant to the same query parameters (i.e., either only certain
parameters are ever meaningful, or certain parameters are always
meaningless).

In such cases, it is possible to avoid searching several stored
responses and take a more efficient approach. This shortcut might be
desirable in, for example, browser disk cache implementations, where the
number of accesses is known to affect cache latency significantly.

Implementations are not required (neither servers nor caches) to take
this shortcut; the worst consequence is an avoidable cache miss.
Incorrect stored responses are never used.

Resolves #2908.