Update requirements on use of Location during upload creation
Previously, there were "mixed" requirements on including a Location
in interim or final response to an upload creation request. This
created some annoying dealing with 104 responses depending on
if they were first or not.

The PR proposes more consistency. If you're going to generate a
Location, it must always be sent and must always be the same value
in all responses. Only a SHOULD requirement is levied on the client
to check these - making it MUST seems a bit too onerous.

Fixes #3183