Clarify when a server MUST supply an upload-length
We have the requirement "if the client knows the representation data's length,
it SHOULD include the `Upload-Length`". Which means the server can be a position
to either know, or not know the value.

This change makes a minor tweak to the text to highlight that when the server
was given a length it really MUST respond with it but when it didn't get given
one it can't.

Fixes #3187