Clarify the idempotence stance in the QUERY method proposal.
Remove the "potentially", to make it clear that it's not safe to
consider POST request idempotent, as contrasted to GET and QUERY.

What matters in practice is not that a POST request can potentially
be idempotent, but that it can potentially not be idempotent. That is,
the guarantee, not a hypothesis. Hence the reformulation.